+++ packages/typespec-azure-core/src/rules/no-multiple-discriminator.ts
@@ -2,12 +2,14 @@ import {
   Discriminator,
   Model,
   createRule,
+  fileRef,
   getDiscriminator,
   paramMessage,
 } from "@typespec/compiler";
 
 export const noMultipleDiscriminatorRule = createRule({
   name: "no-multiple-discriminator",
+  docs: fileRef.fromPackageRoot("src/rules/no-multiple-discriminator.md"),
   description: "Classes should have at most one discriminator.",
   severity: "warning",
   url: "https://azure.github.io/typespec-azure/docs/libraries/azure-core/rules/no-multiple-discriminator",