+++ packages/typespec-azure-core/src/rules/no-rpc-path-params.ts
@@ -1,9 +1,10 @@
-import { Operation, createRule } from "@typespec/compiler";
+import { Operation, createRule, fileRef } from "@typespec/compiler";
 import { isPathParam } from "@typespec/http";
 import { isExcludedCoreType } from "./utils.js";
 
 export const noRpcPathParamsRule = createRule({
   name: "no-rpc-path-params",
+  docs: fileRef.fromPackageRoot("src/rules/no-rpc-path-params.md"),
   description: "Operations defined using RpcOperation should not have path parameters.",
   severity: "warning",
   url: "https://azure.github.io/typespec-azure/docs/libraries/azure-core/rules/no-rpc-path-params",